Bind graph colors
