+++ www/package.json
@@ -60,6 +60,7 @@
     "pluralsh-design-system": "1.165.0",
     "prop-types": "15.8.1",
     "query-string": "7.1.1",
+    "randomcolor": "0.6.2",
     "react": "18.2.0",
     "react-animations": "1.0.0",
     "react-collapsible": "2.10.0",
@@ -108,6 +109,7 @@
     "@babel/preset-react": "7.18.6",
     "@pluralsh/eslint-config-typescript": "2.3.0",
     "@pluralsh/stylelint-config": "1.0.0",
+    "@types/randomcolor": "^0",
     "@typescript-eslint/eslint-plugin": "5.33.1",
     "@typescript-eslint/parser": "5.33.1",
     "eslint": "8.22.0",
